engine/requirements.txt
aniso8601==8.0.0
asn1crypto==0.24.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.1.4
dnspython==1.16.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
gunicorn==20.0.4
httplib2==0.9.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
netifaces==0.10.4
packaging==20.1
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pyflakes==2.1.1
PyJWT==1.7.1
pymongo==3.10.1
PyNaCl==1.3.0
pyparsing==2.4.6
python-dotenv==0.11.0
pytz==2019.3
PyYAML==3.12
redis==3.4.1
requests==2.22.0
rq==1.3.0
six==1.14.0
urllib3==1.25.8
uWSGI==2.0.18
flask_cors==3.0.8
Werkzeug==1.0.0
